Removed unwanted zeros from vault.
Due to the way generate_permitted_abscissas functions it places zeros in
place of NaN or similar type. Due to this the vault has a collection of
zeros that occupy the vault and needed to be removed. This has been done
by constructing a list of points that could occupy the entire vault
without the locking elements. Then remove the excees zeros and then only
insert enough coordinates into the vault until size has been reached.